Rename HgDataTable to HgDataTableComponent
