+++ Apps/WebClient/src/NewClientApp/src/components/common/HgDataTableComponent.vue
