+++ Apps/WebClient/src/NewClientApp/src/components/private/reports/LabResultReportComponent.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { computed, onMounted, watch } from "vue";
 
-import HgDataTable from "@/components/common/HgDataTable.vue";
+import HgDataTableComponent from "@/components/common/HgDataTableComponent.vue";
 import { container } from "@/ioc/container";
 import { SERVICE_IDENTIFIER } from "@/ioc/identifier";
 import { DateWrapper } from "@/models/dateWrapper";
@@ -129,7 +129,7 @@ labResultStore
     <p v-if="isEmpty && !labResultsAreLoading" class="px-4">
         <v-col>No records found.</v-col>
     </p>
-    <HgDataTable
+    <HgDataTableComponent
         v-else-if="!isDependent"
         class="d-none d-md-block"
         fixed-header
